aider: Replace chat widget with a link to the chat in the email template.
# Context:
USER: the `BrevoConversations` chat widget is not included with the sent email message. here is the code for it: `import { useEffect } from "react";

const BrevoConversations = () => {
  useEffect(() => {
    (function (d, w, c) {
      w.BrevoConversationsID = "64ae423dc9574471816653f0";
      w[c] =
        w[c] ||
        function () {
          (w[c].q = w[c].q || []).push(arguments);
        };
      var s = d.createElement("script");
      s.async = true;
      s.src = "https://conversations-widget.brevo.com/brevo-conversations.js";
      if (d.head) d.head.appendChild(s);
    })(document, window, "BrevoConversations");
  }, []);

  return null;
};

export default BrevoConversations;
`
ASSISTANT: The `BrevoConversations` component is a React component that injects a script into the DOM to load the chat widget. This works when the component is rendered in a browser, but it won't work when included in an email because email clients don't execute JavaScript for security reasons.

Instead, you could include a link in the email that opens a webpage with the `BrevoConversations` component. Here's how you can modify the email payload to include this link:

```javascript
frontend/pages/email.js
<<<<<<< ORIGINAL
    chatWidget: `<div>${BrevoConversations()}</div>`, // Include the chat widget in the email template
=======
    chatWidget: `<a href="https://yourwebsite.com/chat">Open Chat</a>`, // Include a link to the chat in the email
>>>>>>> UPDATED
```

Replace `https://yourwebsite.com/chat` with the URL of the page where you have the `BrevoConversations` component.

Please note that you'll need to create a new page that renders the `BrevoConversations` component if you haven't done so already.